+++ test/functional/hooks_test.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:http/http.dart' as http;
 import 'package:json_api/client.dart';
-import 'package:json_api/server.dart';
+import 'package:json_api/url_design.dart';
 import 'package:test/test.dart';
 
 import '../../example/cars_server.dart';